requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
Django==3.0
djangorestframework==3.10.3
idna==2.8
Naked==0.1.31
pbkdf2==1.3
pycrypto==2.6.1
pytz==2019.3
PyYAML==5.2
pyzmq==18.1.1
requests==2.22.0
shellescape==3.4.1
sqlparse==0.3.0
urllib3==1.25.7
zmq==0.0.0
